+++ lib/punchblock/event/unjoined.rb
@@ -6,7 +6,7 @@ module Punchblock
       register :unjoined, :core
 
       # @return [String] the call ID that was unjoined
-      attribute :call_id
+      attribute :call_uri
 
       # @return [String] the mixer name that was unjoined
       attribute :mixer_name