+++ spec/punchblock/ref_spec.rb
@@ -9,7 +9,7 @@ module Punchblock
     end
 
     describe "from a stanza" do
-      let(:stanza) { "<ref id='fgh4590' xmlns='urn:xmpp:rayo:1' />" }
+      let(:stanza) { "<ref uri='fgh4590' xmlns='urn:xmpp:rayo:1' />" }
 
       subject { RayoNode.from_xml parse_stanza(stanza).root, '9f00061', '1' }
 
@@ -17,19 +17,19 @@ module Punchblock
 
       it_should_behave_like 'event'
 
-      its(:id) { should be == 'fgh4590' }
+      its(:uri) { should be == 'fgh4590' }
     end
 
     describe "when setting options in initializer" do
-      subject { Ref.new :id => 'foo' }
+      subject { Ref.new uri: 'foo' }
 
-      its(:id) { should be == 'foo' }
+      its(:uri) { should be == 'foo' }
 
       describe "exporting to Rayo" do
         it "should export to XML that can be understood by its parser" do
           new_instance = RayoNode.from_xml subject.to_rayo
           new_instance.should be_instance_of described_class
-          new_instance.id.should == 'foo'
+          new_instance.uri.should == 'foo'
         end
 
         it "should render to a parent node if supplied" do
@@ -44,7 +44,7 @@ module Punchblock
           subject { described_class.new }
 
           it "should not include them in the XML representation" do
-            subject.to_rayo['id'].should be_nil
+            subject.to_rayo['uri'].should be_nil
           end
         end
       end