+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # [develop](https://github.com/adhearsion/punchblock)
+  * Feature: Compliance with v0.2 of the published Rayo spec (http://xmpp.org/extensions/xep-0327.html)
+  * Feature: Add support for Rayo Prompt component (no support on FS)
   * Change: Models are now plain ruby objects, not XML nodes, and are imported from/exported to XML when necessary for communicating over XMPP.
   * Change: `#headers` and AMI `#attributes` now do not have their names modified. A header of `'Call-ID'` will no longer be modified to `:call_id`.
   * Change: AMI Events/Actions now have `#headers(=)` rather than `#attributes(=)`
@@ -28,6 +30,9 @@
   * Bugfix: Avoid DTMF recognizer failures and race conditions by bringing DTMFRecognizer back into the Input component actor.
   * Bugfix: Catch Asterisk AMI errors in all cases and fail accordingly, instead of ploughing ahead in the face of adversity.
   * Bugfix: Improve performance of Asterisk implementation by no longer spinning up a component actor for AGI command execution.
+
+# [v1.8.2](https://github.com/adhearsion/punchblock/compare/v1.8.1...v1.8.2) - [2013-04-19](https://rubygems.org/gems/punchblock/versions/1.8.2)
+  * Bugfix: Input initial timeout was being set as a float rather than an integer
 
 # [v1.8.2](https://github.com/adhearsion/punchblock/compare/v1.8.1...v1.8.2) - [2013-04-19](https://rubygems.org/gems/punchblock/versions/1.8.2)
   * Bugfix: Input initial timeout was being set as a float rather than an integer